PC is passing! had to shift the expected values of the PC register
back by one clock cycles so it matches.  Best guess to why this is
needed: the way bluespec samples is not the same as how PC unit's
cycle line is set to sample.  It looks like bluespec is sampling when
the clock cycle is low.  I don't know how to change that.